% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_fit.R
\name{logLik_mnl}
\alias{logLik_mnl}
\title{Log-likelihood for a multinomial logit (MNL) model}
\usage{
logLik_mnl(Beta, Data, X, choiceVar, GroupID)
}
\arguments{
\item{Beta}{Numeric vector of parameters.}

\item{Data}{Data frame or data.table in long format.}

\item{X}{Character vector of column names for explanatory variables.}

\item{choiceVar}{Name of the binary choice indicator variable.}

\item{GroupID}{Vector of group identifiers (e.g. individual or choice set).}
}
\value{
Vector of individual log-likelihood contributions.
}
\description{
Log-likelihood for a multinomial logit (MNL) model
}
\keyword{internal}
